+++ packages/stateful_service/lib/stateful_service.dart
@@ -1,6 +1,6 @@
 import 'dart:async';
 
-import 'package:logger/logger.dart';
+import 'package:logging/logging.dart';
 import 'package:meta/meta.dart';
 import 'package:synchronized/synchronized.dart';
 
@@ -45,25 +45,25 @@ abstract class StatefulService<S> {
     bool Function(S previousState, S newState)? shouldStateBeEmitted,
   })  : _state = initialState,
         _name = name,
-        _logger = logger ?? Logger(),
         _cache = cache,
         _shouldStateBeEmitted = shouldStateBeEmitted ?? ((a, b) => a != b) {
+    _logger = logger ?? Logger(runtimeType.toString());
     final cache = _cache;
     if (cache == null) {
       initComplete = Future.value(null);
     } else {
       initComplete = _update(() async {
         final cachedState = await cache.init().onError((err, trace) {
-          _logger.e(
+          _logger.severe(
             '[$name] Failed to initialize ${cache.runtimeType}',
-            error: err,
-            stackTrace: trace,
+            err,
+            trace,
           );
           return null;
         });
         if (cachedState != null && cacheValidator?.call(cachedState) != false) {
           await _addState(cachedState);
-          _logger.d('[$name] State cache initialized');
+          _logger.fine('[$name] State cache initialized');
         } else {
           await cache.put(initialState);
         }
@@ -73,7 +73,7 @@ abstract class StatefulService<S> {
 
   S _state;
   final String? _name;
-  final Logger _logger;
+  late final Logger _logger;
   final StreamController<S> _controller = StreamController.broadcast();
   final StatefulServiceCache<S>? _cache;
   final bool Function(S state1, S state2) _shouldStateBeEmitted;
@@ -116,22 +116,30 @@ abstract class StatefulService<S> {
       _update(() async => _addState(await update(state)), ignoreConcurrentUpdates);
 
   /// Performs a series of updates of service's state with the values emitted from the stream returned by [updates]. The
-  /// service state will be locked (no other updates can take place) until the stream completes.
+  /// service state will be locked (no other updates can take place) until the stream completes. [update] receives two
+  /// arguments; the service's current state, and a function to call to save the current state as a savepoint.
   ///
-  /// If the stream emits an error at any point, the service's state will be reverted to the value it had before the
-  /// execution of the [updates] function started.
+  /// If the stream emits an error at any point, the service's state will be reverted to the the last save point or, if
+  /// no save-call has been made, the state it had before the execution of the [updates] function started.
   ///
   /// If [ignoreConcurrentUpdates] is true, any other update requests will be dropped, rather than queued, until the
   /// [Stream] returned by [updates] has completed.
   @protected
   @visibleForTesting
-  Future<void> streamUpdates(Stream<S> Function(S state) updates, {bool ignoreConcurrentUpdates = false}) =>
+  Future<void> streamUpdates(
+    Stream<S> Function(S state, void Function() save) updates, {
+    bool ignoreConcurrentUpdates = false,
+  }) =>
       _update(() async {
-        final previous = state;
+        var savePoint = state;
+        var previous = savePoint;
         try {
-          await updates(previous).forEach(_addState);
+          await updates(savePoint, () => savePoint = previous).forEach((state) {
+            previous = state;
+            _addState(state);
+          });
         } catch (_) {
-          await _addState(previous);
+          await _addState(savePoint);
           rethrow;
         }
       }, ignoreConcurrentUpdates);
@@ -141,7 +149,7 @@ abstract class StatefulService<S> {
 
     final ignoreUpdates = _ignoreUpdates;
     if (ignoreUpdates) {
-      _logger.d('[$name] Ignoring concurrent state update');
+      _logger.fine('[$name] Ignoring concurrent state update');
       return Future.value(null);
     } else if (ignoreConcurrentUpdates) {
       _ignoreUpdates = true;
@@ -149,7 +157,7 @@ abstract class StatefulService<S> {
     _isUpdating = true;
     return _lock.synchronized(() async {
       if (isClosed) {
-        _logger.w('[$name] Already closed, ignoring update');
+        _logger.warning('[$name] Already closed, ignoring update');
       } else {
         await f();
       }
@@ -165,9 +173,9 @@ abstract class StatefulService<S> {
       if (cache != null) {
         try {
           await cache.put(state);
-          _logger.d('[$name] State written to cache');
+          _logger.fine('[$name] State written to cache');
         } catch (err, trace) {
-          _logger.e('[$name] Failed to write to ${_cache.runtimeType}', error: err, stackTrace: trace);
+          _logger.severe('[$name] Failed to write to ${_cache.runtimeType}', err, trace);
         }
       }
       _state = state;
@@ -181,7 +189,7 @@ abstract class StatefulService<S> {
   @mustCallSuper
   Future<void> close() async {
     await initComplete;
-    _logger.d('[$name] Closing');
+    _logger.fine('[$name] Closing');
     await _controller.close();
     await _cache?.close();
   }